requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
python-multipart==0.0.20
pdfplumber==0.11.7
llama-cpp-python>=0.3.16

